// ui/src/generated/OpenResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DocPayload } from "./DocPayload";
import type { OpenError } from "./OpenError";

export type OpenResult = { "status": "ok", doc: DocPayload, } | { "status": "err", error: OpenError, };