// ui/src/generated/RenderedDoc.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Frontmatter } from "./Frontmatter";
import type { OutlineItem } from "./OutlineItem";
import type { TaskStats } from "./TaskStats";

export type RenderedDoc = { html: string, outline: Array<OutlineItem>, frontmatter: Frontmatter | null, tasks: TaskStats, title: string, wordCount: number, hasUnresolvedWikilinks: boolean, };